+++ src/server/logic/playingfield.h
@@ -7,19 +7,19 @@
 #define CELL(field, x, y) (field)->cell[(y) * (field)->width + (x)]
 
 typedef struct PlayingField {
-    uint16_t height;
-    uint16_t width;
+    int height;
+    int width;
     uint8_t *cell;
 } PlayingField;
 
-int init_playingField(PlayingField *field, uint16_t w, uint16_t h);
+int init_playingField(PlayingField *field, uint8_t w, uint8_t h);
 void print_playingField(PlayingField *field);
 void free_playingField(PlayingField *field);
 void prepare_playingField(PlayingField *field);
 
-uint8_t SAFE_GET_CELL(PlayingField* field, uint16_t x, uint16_t y);
-uint8_t SAFE_SET_CELL(PlayingField* field, uint16_t x, uint16_t y, uint8_t v);
-uint8_t move_cell_contents(PlayingField* field, uint16_t x1, uint16_t y1, uint16_t x2, uint16_t y2);
-uint16_t cell_to_uint(PlayingField* field, uint16_t x, uint16_t y);
+uint8_t SAFE_GET_CELL(PlayingField* field, uint8_t x, uint8_t y);
+uint8_t SAFE_SET_CELL(PlayingField* field, uint8_t x, uint8_t y, uint8_t v);
+uint8_t move_cell_contents(PlayingField* field, uint8_t x1, uint8_t y1, uint8_t x2, uint8_t y2);
+uint8_t cell_to_uint(PlayingField* field, uint8_t x, uint8_t y);
 
 #endif
